Remove base id argument from `mkBaseUrl`
Base specific APIs need a base identifier while constructing the base URL but
since we're concerning ourselves with the Metadata API for the time being we
don't need it sticking around. If later on we want to share the utilities then
we can supply it as a `Maybe` field and have two different environments.